requirements.txt
Django==3.2.12
django-mptt==0.12.0